requirements.txt
# requirements.txt - UNPINNED FOR MODERN RESOLVER

# --- Core AI Agent Framework ---
crewai
crewai-tools

# --- Frontend Web Application ---
streamlit

# --- Integration Layer for Tools ---
composio-crewai

# --- LLM and LanguageChain Support ---
langchain-google-genai
google-generativeai
# openai is often a required dependency for underlying tools, so we keep it
openai

# --- Utilities ---
python-dotenv
requests
Pillow